move functions to lexos/interfaces
