+++ lexos/managers/utility.py
@@ -237,7 +237,7 @@ def generate_csv(file_manager: FileManager) -> Tuple[str, str]:
     return out_file_path, extension
 
 
-# Gets called from statistics() in lexos.py
+# Gets called from statistics() in lexos_core.py
 
 
 def generate_statistics(file_manager: FileManager) -> \
@@ -398,7 +398,7 @@ def get_newick(node, newick, parent_dist, leaf_names):
         return newick
 
 
-# Gets called from cluster() in lexos.py
+# Gets called from cluster() in lexos_core.py
 def generate_dendrogram(file_manager: FileManager, leq: str):
     """
     Generates dendrogram image and PDF from the active files.
@@ -685,7 +685,7 @@ def generate_k_means_pca(file_manager: FileManager):
     return kmeans_index, siltt_score, file_name_str, k_value, color_chart
 
 
-# Gets called from kmeans() in lexos.py
+# Gets called from kmeans() in lexos_core.py
 
 
 def generate_k_means_voronoi(file_manager: FileManager):
@@ -1741,7 +1741,7 @@ def xml_handling_options(data: dict = {}):
             del session_manager.session['xmlhandlingoptions'][key]
 
 
-# Gets called from cluster() in lexos.py
+# Gets called from cluster() in lexos_core.py
 def generate_dendrogram_from_ajax(file_manager: FileManager, leq: str):
     """
     Generates dendrogram image and PDF from the active files.
